Move image and data filter to ImageHelper
